make chat more real time
